requirements.txt
pillow == 4.1.1
xmltodict == 0.11.0
numpy == 1.13.0
matplotlib == 2.0.2
shapely == 1.5.13
fiona == 1.7.8
